add ClientErrorException.ProhibitedActionException to handle spring security exceptions update BrandServiceImpl.java add templates for admin adjust DataInitializer.java add forms

+++ src/main/java/com/example/carcatalog/controller/OfferController.java
@@ -1,7 +1,12 @@
 package com.example.carcatalog.controller;
 
+import com.example.carcatalog.dto.ModelDTO;
+import com.example.carcatalog.dto.OfferDTO;
 import com.example.carcatalog.dto.viewmodel.OfferViewModel;
+import com.example.carcatalog.except.ClientErrorException;
+import com.example.carcatalog.service.ModelService;
 import com.example.carcatalog.service.OfferService;
+import com.example.carcatalog.service.UserService;
 import com.example.carcatalog.service.impl.RecentOfferService;
 import lombok.extern.java.Log;
 import org.apache.logging.log4j.Level;
@@ -10,11 +15,11 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.*;
 
 import java.security.Principal;
+import java.util.List;
 import java.util.UUID;
 
 /**
@@ -31,11 +36,16 @@ import java.util.UUID;
 public class OfferController {
     private RecentOfferService recentOfferService;
     private OfferService offerService;
-
+    private ModelService modelService;
 
     @Autowired
     public void setOfferService(OfferService offerService) {
         this.offerService = offerService;
+    }
+
+    @Autowired
+    public void setModelService(ModelService modelService) {
+        this.modelService = modelService;
     }
 
     @Autowired
@@ -74,14 +84,71 @@ public class OfferController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteOffer(@PathVariable("id") UUID offerId, Model model) {
-        offerService.delete(offerId);
-        return "redirect:/";
+    public String deleteOffer(@PathVariable("id") UUID offerId, Principal principal) {
+
+        if (offerService.findById(offerId).getSellerUsername().equals(principal.getName())) {
+            offerService.delete(offerId);
+            return "redirect:/users/" + principal.getName();
+        }
+
+        throw new ClientErrorException.ProhibitedActionException("delete", "offer");
     }
 
     @GetMapping("/recent")
     public String getRecent(Model model, Principal principal) {
         model.addAttribute("offers", recentOfferService.getRecent(principal.getName()));
         return "home";
+    }
+
+    @GetMapping("/create-offer")
+    public String showCreateOfferForm(Model model) {
+        List<ModelDTO> availableModels = modelService.findAll();
+        model.addAttribute("availableModels", availableModels);
+
+        model.addAttribute("offer", new OfferDTO());
+        return "create-offer";
+    }
+
+    @PostMapping("/create-offer")
+    public String processCreateOfferForm(@ModelAttribute("offer") OfferDTO offer,
+                                         BindingResult bindingResult,
+                                         Principal principal) {
+        if (bindingResult.hasErrors()) {
+            return "create-offer";
+        }
+
+        offer.setSellerUsername(principal.getName());
+        offerService.add(offer);
+        return "redirect:/users/dashboard";
+    }
+
+    @GetMapping("/edit/{id}")
+    public String showEditOfferForm(@PathVariable("id") UUID offerId, Model model) {
+        OfferDTO offerDTO = offerService.findById(offerId);
+
+        List<ModelDTO> availableModels = modelService.findAll();
+        model.addAttribute("availableModels", availableModels);
+
+        model.addAttribute("offer", offerDTO);
+        return "edit-offer";
+    }
+
+    @PostMapping("/edit/{id}")
+    public String processEditOfferForm(@PathVariable("id") UUID offerId,
+                                       @ModelAttribute("offer") OfferDTO offer,
+                                       BindingResult bindingResult,
+                                       Principal principal) {
+        if (!offerService.findById(offerId).getSellerUsername().equals(principal.getName())) {
+            throw new ClientErrorException.ProhibitedActionException("edit", "offer");
+        }
+
+        if (bindingResult.hasErrors()) {
+            return "edit-offer";
+        }
+
+        offer.setSellerUsername(principal.getName());
+        offer.setId(offerId);
+        offerService.update(offer);
+        return "redirect:/users/dashboard";
     }
 }
